Apply escaping args to other command shells
